src/main/java/subway/application/SectionService.java
package subway.application;

import org.springframework.stereotype.Service;
import subway.dao.SectionDao;
import subway.domain.Section;
import subway.dto.SectionRequest;
import subway.dto.SectionResponse;

@Service
public class SectionService {

    private final SectionDao sectionDao;

    public SectionService(SectionDao sectionDao) {
        this.sectionDao = sectionDao;
    }

    public SectionResponse saveSection(Long lineId, SectionRequest request) {
        Section section = sectionDao.insert(request.toEntity(lineId));
        return SectionResponse.from(section);
    }
}
